requirements.txt
pycti==3.3.3
PyYAML==5.3.1
pika==1.1.0
art==4.8
progressbar2==3.53.1